Split resolver expression validation
