+++ src/resolver/local_validation.rs
@@ -1,0 +1,193 @@
+use crate::ast::{Param, Pattern, Statement, TypeParam};
+use crate::error::{Diagnostic, Span};
+
+use super::symbol_table::ScopeStack;
+use super::{Namespace, Resolver, SymbolTable};
+
+impl Resolver {
+    pub(super) fn validate_statement_refs(
+        &self,
+        table: &mut SymbolTable,
+        type_params: &[TypeParam],
+        statement: &Statement,
+        locals: &mut ScopeStack,
+        allow_self_type: bool,
+        diagnostics: &mut Vec<Diagnostic>,
+    ) {
+        match statement {
+            Statement::VarDecl {
+                name,
+                ty,
+                value,
+                mutable,
+                constant,
+                ..
+            } => {
+                if let Some(ty) = ty {
+                    self.validate_type_ref(
+                        table,
+                        type_params,
+                        ty,
+                        statement.span(),
+                        allow_self_type,
+                        diagnostics,
+                    );
+                }
+                self.validate_expr_refs(
+                    table,
+                    type_params,
+                    value,
+                    locals,
+                    allow_self_type,
+                    diagnostics,
+                );
+                if *constant || *mutable || !locals.is_mutable(name) {
+                    self.define_local_symbol(
+                        table,
+                        name,
+                        *mutable,
+                        statement.span(),
+                        locals,
+                        diagnostics,
+                    );
+                }
+            }
+            Statement::Assignment { target, value, .. } => {
+                self.validate_expr_refs(
+                    table,
+                    type_params,
+                    target,
+                    locals,
+                    allow_self_type,
+                    diagnostics,
+                );
+                self.validate_expr_refs(
+                    table,
+                    type_params,
+                    value,
+                    locals,
+                    allow_self_type,
+                    diagnostics,
+                );
+            }
+            Statement::Expression { expr, .. } => {
+                self.validate_expr_refs(
+                    table,
+                    type_params,
+                    expr,
+                    locals,
+                    allow_self_type,
+                    diagnostics,
+                );
+            }
+            Statement::Block { stmts, .. } => {
+                let block_scope_id = table.new_scope();
+                let mut block_locals = ScopeStack::with_parent(block_scope_id, locals);
+                for statement in stmts {
+                    self.validate_statement_refs(
+                        table,
+                        type_params,
+                        statement,
+                        &mut block_locals,
+                        allow_self_type,
+                        diagnostics,
+                    );
+                }
+            }
+        }
+    }
+
+    pub(super) fn is_known_value_name(
+        &self,
+        table: &SymbolTable,
+        locals: &ScopeStack,
+        name: &str,
+    ) -> bool {
+        table.lookup(Namespace::Value, name).is_some()
+            || table.lookup(Namespace::Import, name).is_some()
+            || locals.contains(name)
+    }
+
+    pub(super) fn param_locals(
+        &self,
+        table: &mut SymbolTable,
+        params: &[Param],
+        scope_id: u32,
+        diagnostics: &mut Vec<Diagnostic>,
+    ) -> ScopeStack {
+        let mut locals = ScopeStack::new(scope_id);
+        for param in params {
+            self.define_local_symbol(
+                table,
+                &param.name,
+                param.mutable,
+                param.span,
+                &mut locals,
+                diagnostics,
+            );
+        }
+        locals
+    }
+
+    pub(super) fn define_local_symbol(
+        &self,
+        table: &mut SymbolTable,
+        name: &str,
+        mutable: bool,
+        span: Span,
+        locals: &mut ScopeStack,
+        diagnostics: &mut Vec<Diagnostic>,
+    ) {
+        match table.define_local(name, mutable, locals.current_scope_id, span) {
+            Ok(_) => locals.insert(name.to_string(), mutable),
+            Err(diagnostic) => diagnostics.push(*diagnostic),
+        }
+    }
+
+    pub(super) fn bind_pattern_locals(
+        &self,
+        table: &mut SymbolTable,
+        pattern: &Pattern,
+        locals: &mut ScopeStack,
+        diagnostics: &mut Vec<Diagnostic>,
+    ) {
+        match pattern {
+            Pattern::Identifier { name, span } => {
+                self.define_local_symbol(table, name, false, *span, locals, diagnostics);
+            }
+            Pattern::Struct { fields, .. } => {
+                for (name, nested) in fields {
+                    if let Some(nested) = nested {
+                        self.bind_pattern_locals(table, nested, locals, diagnostics);
+                    } else {
+                        self.define_local_symbol(
+                            table,
+                            name,
+                            false,
+                            pattern.span(),
+                            locals,
+                            diagnostics,
+                        );
+                    }
+                }
+            }
+            Pattern::Enum {
+                payload: Some(payload),
+                ..
+            } => {
+                self.bind_pattern_locals(table, payload, locals, diagnostics);
+            }
+            Pattern::Or { patterns, .. } => {
+                for pattern in patterns {
+                    self.bind_pattern_locals(table, pattern, locals, diagnostics);
+                }
+            }
+            Pattern::Wildcard { .. }
+            | Pattern::Literal { .. }
+            | Pattern::Enum { payload: None, .. }
+            | Pattern::Range { .. }
+            | Pattern::BoolTrue { .. }
+            | Pattern::BoolFalse { .. } => {}
+        }
+    }
+}
